Fixing major bug
Permutations weren't using the permutation code in the cross-validation, this meant I was running about 100x as many permutations as necessary. Also I was calculating permutations for all four spatial/feature runs, but you really only need to do it for one of the four. Fixes should reduce # of function calls by a lot and uses a faster fitting routine.